// src/app/Models/companies/company.ts
export class Company {
    IdCompany: number;
    CompanyName: string;
    Code: string;
    Description: string;
    Direction: string;
    City: string;
    Department: string;
    ImageURL: string;
    State: boolean;
    UserCreate: string;
    CreateDate: Date;
    UserModified: string;
    ModifiedDate: Date;
}